requirements.txt
fastapi==0.128.0
langchain_community==0.4.1
langchain_core==1.2.5
langchain_openai==1.1.6
langchain_qdrant==1.1.0
langchain_text_splitters==1.1.0
langgraph==1.0.5
pydantic==2.12.5
python-dotenv==1.2.1
qdrant_client==1.16.2
Requests==2.32.5
vllm==0.13.0